move onclick hook to hooks
